Extract Storm Setup ingestion responsibilities #298
